+++ client/driver/utils.go
@@ -1,7 +1,6 @@
 package driver
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -9,10 +8,8 @@ import (
 	"os/exec"
 	"path/filepath"
 	"strings"
-	"syscall"
 	"time"
 
-	"github.com/armon/circbuf"
 	"github.com/hashicorp/go-multierror"
 	"github.com/hashicorp/go-plugin"
 	"github.com/hashicorp/nomad/client/config"
@@ -180,37 +177,4 @@ func getExecutorUser(task *structs.Task) string {
 		return cstructs.DefaultUnpriviledgedUser
 	}
 	return task.User
-}
-
-// execChroot executes cmd with args inside chroot if set and returns the
-// output, exit code, and error. If chroot is an empty string the command is
-// executed on the host.
-func execChroot(ctx context.Context, chroot, name string, args []string) ([]byte, int, error) {
-	buf, _ := circbuf.NewBuffer(int64(cstructs.CheckBufSize))
-	cmd := exec.CommandContext(ctx, name, args...)
-	cmd.Dir = "/"
-	cmd.Stdout = buf
-	cmd.Stderr = buf
-	if chroot != "" {
-		setChroot(cmd, chroot)
-	}
-	if err := cmd.Run(); err != nil {
-		exitErr, ok := err.(*exec.ExitError)
-		if !ok {
-			// Non-exit error, return it and let the caller treat
-			// it as a critical failure
-			return nil, 0, err
-		}
-
-		// Some kind of error happened; default to critical
-		exitCode := 2
-		if status, ok := exitErr.Sys().(syscall.WaitStatus); ok {
-			exitCode = status.ExitStatus()
-		}
-
-		// Don't return the exitError as the caller only needs the
-		// output and code.
-		return buf.Bytes(), exitCode, nil
-	}
-	return buf.Bytes(), 0, nil
 }