+++ client/task_runner_test.go
@@ -75,6 +75,7 @@ func testTaskRunner(t *testing.T, restarts bool) *taskRunnerTestCtx {
 func testTaskRunnerFromAlloc(t *testing.T, restarts bool, alloc *structs.Allocation) *taskRunnerTestCtx {
 	logger := testLogger()
 	conf := config.DefaultConfig()
+	conf.Node = mock.Node()
 	conf.StateDir = os.TempDir()
 	conf.AllocDir = os.TempDir()
 
@@ -380,7 +381,7 @@ func TestTaskRunner_SaveRestoreState(t *testing.T) {
 	tr2 := NewTaskRunner(ctx.tr.logger, ctx.tr.config, ctx.tr.stateDB, ctx.upd.Update,
 		ctx.tr.taskDir, ctx.tr.alloc, task2, ctx.tr.vaultClient, ctx.tr.consul)
 	tr2.restartTracker = noRestartsTracker()
-	if err := tr2.RestoreState(); err != nil {
+	if _, err := tr2.RestoreState(); err != nil {
 		t.Fatalf("err: %v", err)
 	}
 	go tr2.Run()
@@ -1530,4 +1531,50 @@ func TestTaskRunner_CleanupFail(t *testing.T) {
 	if !reflect.DeepEqual(expected, ctx.tr.createdResources.Resources) {
 		t.Fatalf("expected %#v but found: %#v", expected, ctx.tr.createdResources.Resources)
 	}
+}
+
+func TestTaskRunner_Pre06ScriptCheck(t *testing.T) {
+	run := func(ver, driver, checkType string, exp bool) (string, func(t *testing.T)) {
+		name := fmt.Sprintf("%s %s %s returns %t", ver, driver, checkType, exp)
+		return name, func(t *testing.T) {
+			services := []*structs.Service{
+				{
+					Checks: []*structs.ServiceCheck{
+						{
+							Type: checkType,
+						},
+					},
+				},
+			}
+			if act := pre06ScriptCheck(ver, driver, services); act != exp {
+				t.Errorf("expected %t received %t", exp, act)
+			}
+		}
+	}
+	t.Run(run("0.5.6", "exec", "script", true))
+	t.Run(run("0.5.6", "java", "script", true))
+	t.Run(run("0.5.6", "mock_driver", "script", true))
+	t.Run(run("0.5.9", "exec", "script", true))
+	t.Run(run("0.5.9", "java", "script", true))
+	t.Run(run("0.5.9", "mock_driver", "script", true))
+
+	t.Run(run("0.6.0dev", "exec", "script", false))
+	t.Run(run("0.6.0dev", "java", "script", false))
+	t.Run(run("0.6.0dev", "mock_driver", "script", false))
+	t.Run(run("0.6.0", "exec", "script", false))
+	t.Run(run("0.6.0", "java", "script", false))
+	t.Run(run("0.6.0", "mock_driver", "script", false))
+	t.Run(run("1.0.0", "exec", "script", false))
+	t.Run(run("1.0.0", "java", "script", false))
+	t.Run(run("1.0.0", "mock_driver", "script", false))
+
+	t.Run(run("0.5.6", "rkt", "script", false))
+	t.Run(run("0.5.6", "docker", "script", false))
+	t.Run(run("0.5.6", "qemu", "script", false))
+	t.Run(run("0.5.6", "raw_exec", "script", false))
+	t.Run(run("0.5.6", "invalid", "script", false))
+
+	t.Run(run("0.5.6", "exec", "tcp", false))
+	t.Run(run("0.5.6", "java", "tcp", false))
+	t.Run(run("0.5.6", "mock_driver", "tcp", false))
 }